refactor(core): 持久后端 helper 收编进 persistent-backend.ts 消除两处同构重复
review 中发现 worker-pool 新增的 resolveBackendTypeForKill + 三路 switch 与
session-manager 既有的 getSessionPersistentBackendType / persistentSessionName /
probePersistentSession / killPersistentSession（PR#167 引入）高度同构。抽到新模块
src/core/persistent-backend.ts 统一维护（该模块保持零重依赖——worker-pool 与
session-manager 之间本就有 import 环，共享代码不能再挂在任何一边）。

顺带两点语义收益：
- getSessionPersistentBackendType 统一为 initConfig.backendType 优先（per-session
  spawn 时真值，挂起期间仍在 ds 上、能跟住 bot 配置漂移）→ bot 配置 → daemon 默认。
  session-manager 的恢复 sweep / 终端唤醒也吃到这个精度；恢复路径 ds 尚无
  initConfig，行为不变。
- isSuspendableBackendType 移入并升级为 type predicate，调用侧不再需要二次窄化。
